Apply LPA findings: internalize FilterCompiler/ResolvedEventExtensions/EventLogDataQueryExtensions/ProviderSource, seal Provider models (EventModel/MessageModel/ProviderDetails), drop Runtime IVT from DatabaseTools, move 4 GetEventValues tests from Runtime.Tests to Filtering.Tests, refresh FilterCompiler XML doc post-internalize
